refactor: create competitions module
