+++ src/modules/competitions/types.ts
